% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_carto.R
\docType{data}
\name{carto_com}
\alias{carto_com}
\title{carto_com}
\format{
A data frame with 1292 rows and 16 variables:
\describe{
  \item{ DEPCOM }{ code commune, factor }
  \item{ AREA }{ surface de la commune, units }
  \item{ geometry }{ contours de la commune, sfc_MULTIPOLYGON,sfc }
  \item{ Zone }{ Libelle de la commune, factor }
  \item{ CONSO_GWH }{ consommation electrique annuelle de la commune en GWh, numeric }
  \item{ EPCI }{ EPCI auquel appartient la commune, factor }
  \item{ prod_GWh }{ production electrique renouvelable annuelle de la commune, numeric }
  \item{ pourcent_enrr }{ part de la consommation electrique couverte par la production renouvelable, numeric }
  \item{ cat_prct_enrr }{ classe de l'indicateur part de la consommation electrique couverte par la production renouvelable, factor }
  \item{ injections_GWh }{ injections annuelles de biomethane de la commune, numeric }
  \item{ conso_gaz_GWh }{ consommation de gaz annuelle de la commune en GWh, numeric }
  \item{ pourcent_bioch4 }{ part de la consommation de gaz couverte par la production renouvelable, numeric }
  \item{ cat_prct_bioch4 }{ classe de l'indicateur part de la consommation de gaz couverte par la production renouvelable, factor }
  \item{ Sources }{ operateurs electriques competents sur la commune, factor }
  \item{ source_inject }{ source du volume annuel injecte de la commune : GRDF ou estimation DREAL, list }
  \item{ source_conso_gaz }{ operateurs gaziers competents sur la commune, list }
}
}
\source{
DREAL PdL - TEO
}
\usage{
carto_com
}
\description{
Un dataset comprenant la geographie des communes de la region et quelques indicateurs des millesimes les plus recents à cartographier.
}
\keyword{datasets}
