% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_carto.R, R/data_utils_df.R
\docType{data}
\name{carto_dep}
\alias{carto_dep}
\title{carto_dep}
\format{
A data frame with 5 rows and 3 variables:
\describe{
  \item{ DEP }{  factor }
  \item{ NOM_DEP }{  factor }
  \item{ geometry }{  sfc_GEOMETRY,sfc }
}

A data frame with 5 rows and 3 variables:
\describe{
  \item{ DEP }{  factor }
  \item{ NOM_DEP }{  factor }
  \item{ geometry }{  sfc_GEOMETRY,sfc }
}
}
\source{
DREAL PdL - TEO

DREAL PdL - TEO
}
\usage{
carto_dep

carto_dep
}
\description{
Un dataset comprenant la geographie des departements de la region.

Description.
}
\keyword{datasets}
