% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_carto.R
\docType{data}
\name{carto_epci}
\alias{carto_epci}
\title{carto_epci}
\format{
A data frame with 72 rows and 9 variables:
\describe{
  \item{ EPCI }{ code de l'EPCI, factor }
  \item{ AREA }{ surface de l'EPCI, units }
  \item{ geometry }{ contours geographiques de l'EPCI, sfc_GEOMETRY,sfc }
  \item{ Echelle }{ echelle du dataset, factor }
  \item{ Zone }{ libelle de l'EPCI, factor }
  \item{ pourcent_enrr }{ part de la consommation electrique par la production renouvelable, numeric }
  \item{ cat_prct_enrr }{ classe de l'indicateur part de la consommation electrique par la production renouvelable, factor }
  \item{ pourcent_bioch4 }{ part de la consommation de gaz couverte par la production renouvelable numeric }
  \item{ cat_prct_bioch4 }{ classe de l'indicateur part de la consommation de gaz couverte par la production renouvelable, factor }
}
}
\source{
DREAL PdL - TEO
}
\usage{
carto_epci
}
\description{
Un dataset comprenant la geographie des EPCI de la region et quelques indicateurs à cartographier.
}
\keyword{datasets}
