% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_carto.R
\docType{data}
\name{carto_iris}
\alias{carto_iris}
\title{carto_iris}
\format{
A data frame with 2009 rows and 19 variables:
\describe{
  \item{ CODE_IRIS }{ Code IRIS, factor }
  \item{ CONSO_GWH }{ Consommation electrique annuelle de l'IRIS en GWh, numeric }
  \item{ Sources }{ operateurs electriques competents sur l'IRIS, factor }
  \item{ sce_estim }{ origine de la consommation electrique : donnee SDES brute ou estimation DREAL d'un agregat secretise, factor }
  \item{ geometry }{ contours de l'iris, sfc_MULTIPOLYGON,sfc }
  \item{ prod_GWh }{ production electrique renouvelable annuelle de l'IRIS, numeric }
  \item{ pourcent_enrr }{ part de la consommation electrique couverte par la production renouvelable, numeric }
  \item{ cat_prct_enrr }{ classe de l'indicateur part de la consommation electrique couverte par la production renouvelable, factor }
  \item{ NOM_IRIS }{ Nom de l'IRIS, factor }
  \item{ TYP_IRIS }{ Type de l'IRIS : habitat (H), activités économiques (A), divers (D) et commune non irisee (Z) plus d'info sur https://www.insee.fr/fr/information/2438155, factor }
  \item{ DEPCOM }{ code de la commune à laquelle appartient l'IRIS, factor }
  \item{ source_inject }{ source du volume annuel injecte de l'IRIS : GRDF ou estimation DREAL, list }
  \item{ injections_GWh }{ injections annuelles de biomethane de l'IRIS, numeric }
  \item{ conso_gaz_GWh }{ consommation de gaz annuelle de l'IRIS en GWh, numeric }
  \item{ source_conso_gaz }{ operateurs gaziers competents sur l'IRIS, list }
  \item{ pourcent_bioch4 }{ part de la consommation de gaz couverte par la production renouvelable, numeric }
  \item{ cat_prct_bioch4 }{ classe de l'indicateur part de la consommation de gaz couverte par la production renouvelable, factor }
  \item{ EPCI }{ code de l'EPCI d'appartenance de l'IRIS, factor }
  \item{ DEP }{ code du departement d'appartenance de l'IRIS, factor }
}
}
\source{
DREAL PdL - TEO
}
\usage{
carto_iris
}
\description{
Un dataset comprenant la geographie des IRIS de la region et quelques indicateurs à cartographier.
Les IRIS, ou Ilots Regroupes pour l'Information Statistique, sont un zonage infracommunal defini par l'INSEE et l'IGN.
}
\keyword{datasets}
