% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_carto.R
\docType{data}
\name{carto_reg}
\alias{carto_reg}
\title{carto_reg}
\format{
A data frame with 1 rows and 3 variables:
\describe{
  \item{ REG }{  factor }
  \item{ NOM_REG }{  factor }
  \item{ geometry }{  sfc_MULTIPOLYGON,sfc }
}
}
\source{
DREAL PdL - TEO
}
\usage{
carto_reg
}
\description{
Un dataset comprenant la geographie de la region.
}
\keyword{datasets}
