% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_couleurs.R
\docType{data}
\name{col_biogaz}
\alias{col_biogaz}
\title{col_biogaz}
\format{
An object of class \code{character} of length 6.
}
\source{
DREAL PdL
}
\usage{
col_biogaz
}
\description{
Les 6 couleurs des types d'installations biomethane.
}
\keyword{datasets}
