% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_couleurs.R
\docType{data}
\name{col_enedis}
\alias{col_enedis}
\title{col_enedis}
\format{
An object of class \code{character} of length 4.
}
\source{
DREAL PdL
}
\usage{
col_enedis
}
\description{
Les 4 couleurs des filieres electriques Enedis.
}
\keyword{datasets}
