% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_couleurs.R
\docType{data}
\name{col_registre}
\alias{col_registre}
\title{col_registre}
\format{
An object of class \code{character} of length 6.
}
\source{
DREAL PdL
}
\usage{
col_registre
}
\description{
Les 6 couleurs des filieres electriques registre.
}
\keyword{datasets}
