% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_couleurs.R
\docType{data}
\name{col_ter}
\alias{col_ter}
\title{col_ter}
\format{
An object of class \code{character} of length 1.
}
\source{
DREAL PdL
}
\usage{
col_ter
}
\description{
La couleur de mise en avant du territoire selectionne.
}
\keyword{datasets}
