% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_csv.R
\docType{data}
\name{com_iris_csv_biogaz}
\alias{com_iris_csv_biogaz}
\title{com_iris_csv_biogaz}
\format{
A data frame with 415926 rows and 10 variables:
\describe{
  \item{ CodeZone }{  factor }
  \item{ type }{  character }
  \item{ annee }{  character }
  \item{ source }{  list }
  \item{ indicateur }{  character }
  \item{ valeur }{  numeric }
  \item{ Zone }{  factor }
  \item{ EPCI }{  factor }
  \item{ DEP }{  factor }
  \item{ TypeZone }{  character }
}
}
\source{
DREAL PdL - TEO
}
\usage{
com_iris_csv_biogaz
}
\description{
Un dataset au format long contenant tous les indicateurs lies au biogaz, pour toutes les echelles.
}
\keyword{datasets}
