% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_csv.R
\docType{data}
\name{com_iris_csv_elec}
\alias{com_iris_csv_elec}
\title{com_iris_csv_elec}
\format{
A data frame with 182148 rows and 11 variables:
\describe{
  \item{ CodeZone }{  factor }
  \item{ Zone }{  factor }
  \item{ EPCI }{  factor }
  \item{ Filiere.de.production }{  factor }
  \item{ estimation }{  logical }
  \item{ type_estimation }{  factor }
  \item{ annee }{  factor }
  \item{ indicateur }{  factor }
  \item{ valeur }{  numeric }
  \item{ DEP }{  factor }
  \item{ TypeZone }{  character }
}
}
\source{
DREAL PdL - TEO
}
\usage{
com_iris_csv_elec
}
\description{
Un dataset au format long contenant tous les indicateurs lies a l'electricite, pour toutes les echelles.
}
\keyword{datasets}
