% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cartes_inst_elec.R
\name{couche_fil}
\alias{couche_fil}
\title{Carte mapview des installations d'une filiere electrique du territoire selectionne}
\usage{
couche_fil(fil, code_ter, col, leg = FALSE)
}
\arguments{
\item{fil}{un motif de texte a retrouver dans le code typo des installations electrique}

\item{code_ter}{le code geo du territoire a visualiser, au format texte}

\item{col}{un entier entre 1 et 6 qui designe la filiere dans la palette col_registre}
}
\value{
une carte mapview
}
\description{
Carte mapview des installations d'une filiere electrique du territoire selectionne
}
\examples{
couche_fil(fil = "eol", code_ter = "200060010", col = 5)
}
