% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_indicateurs.R
\docType{data}
\name{Enedis_com_a_reg}
\alias{Enedis_com_a_reg}
\title{Enedis_com_a_reg}
\format{
A data frame with 139620 rows and 8 variables:
\describe{
  \item{ TypeZone }{  factor }
  \item{ CodeZone }{  factor }
  \item{ Filiere.de.production }{  factor }
  \item{ annee }{  factor }
  \item{ indicateur }{  factor }
  \item{ valeur }{  numeric }
  \item{ estimation }{  logical }
  \item{ type_estimation }{  factor }
}
}
\source{
DREAL PdL - TEO
}
\usage{
Enedis_com_a_reg
}
\description{
Les indicateurs electricite renouvelable de la source Enedis, de la commune a la region.
}
\keyword{datasets}
