% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compose_carto.R
\name{get_carto}
\alias{get_carto}
\title{get_carto recupérer une couche pa}
\usage{
get_carto(list, txt = "carte_PV")
}
\arguments{
\item{list}{la liste d'objets reactifs comprenant les différentes couches mapview des filieres}

\item{txt}{le nom d'une couche carto, au format texte}
}
\value{
l'objet mapview correspondant
}
\description{
get_carto recupérer une couche pa
}
\examples{
\dontrun{
get_carto(list = r, "carte_PV")
}
}
