% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_girafeTEO.R
\name{girafeTEO}
\alias{girafeTEO}
\title{Passage d'un graphique ggplot a un graphique ggirafe interactif avec options}
\usage{
girafeTEO(plot, fill_tooltip = TRUE)
}
\arguments{
\item{plot}{l'objet graphique ggplot}

\item{fill_tooltip}{boleen qui indique si on souhaite reprendre la couleur fill du graphique dans le tooltip}
}
\value{
un htmlwidget ggiraph
}
\description{
Passage d'un graphique ggplot a un graphique ggiraph interactif avec options de mise en page
}
\examples{
library(ggplot2)
girafeTEO(ggplot(mpg, aes(hwy)) + geom_bar(), FALSE)
}
