% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_indicateurs.R
\docType{data}
\name{indic_biogaz_epci_a_reg_reg}
\alias{indic_biogaz_epci_a_reg_reg}
\title{indic_biogaz_epci_a_reg_reg}
\format{
A data frame with 9828 rows and 8 variables:
\describe{
  \item{ CodeZone }{  factor }
  \item{ Zone }{  factor }
  \item{ type }{  factor }
  \item{ annee }{  factor }
  \item{ source }{  list }
  \item{ TypeZone }{  factor }
  \item{ indicateur }{  character }
  \item{ valeur }{  numeric }
}
}
\source{
DREAL PdL - TEO
}
\usage{
indic_biogaz_epci_a_reg_reg
}
\description{
Les indicateurs biomethane, de l'epci a la region.
}
\keyword{datasets}
