% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_indicateurs.R
\docType{data}
\name{indic_epci_a_reg}
\alias{indic_epci_a_reg}
\title{indic_epci_a_reg}
\format{
A data frame with 78 rows and 12 variables:
\describe{
  \item{ TypeZone }{  factor }
  \item{ CodeZone }{  factor }
  \item{ Zone }{  factor }
  \item{ prod_GWh }{  numeric }
  \item{ CONSO_GWH }{  numeric }
  \item{ pourcent_enrr }{  numeric }
  \item{ injections_GWh }{  numeric }
  \item{ conso_gaz_GWh }{  numeric }
  \item{ pourcent_bioch4 }{  numeric }
  \item{ source_inject }{  list }
  \item{ source_conso_gaz }{  list }
  \item{ Sources }{  factor }
}
}
\source{
DREAL PdL - TEO
}
\usage{
indic_epci_a_reg
}
\description{
Indicateurs de production/consommation de l'epci à la region, au format large.
}
\keyword{datasets}
