% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_indicateurs.R
\docType{data}
\name{indic_registre}
\alias{indic_registre}
\title{indic_registre}
\format{
A data frame with 62810 rows and 4 variables:
\describe{
  \item{ CodeZone }{  factor }
  \item{ variable }{  factor }
  \item{ valeur }{  numeric }
  \item{ TypeZone }{  factor }
}
}
\source{
DREAL PdL - TEO
}
\usage{
indic_registre
}
\description{
Les indicateurs electricite renouvelable de la source registre, de la commune a la region.
}
\keyword{datasets}
