% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_installations.R
\docType{data}
\name{inst_reg}
\alias{inst_reg}
\title{inst_reg}
\format{
A data frame with 2912 rows and 23 variables:
\describe{
  \item{ nominstallation }{  character }
  \item{ DEPCOM }{  character }
  \item{ IRIS }{  character }
  \item{ date_inst }{  Date }
  \item{ puiss_MW }{  numeric }
  \item{ code_typo }{  character }
  \item{ typo }{  factor }
  \item{ combustible }{  character }
  \item{ combustiblessecondaires }{  character }
  \item{ technologie }{  character }
  \item{ nbgroupes }{  numeric }
  \item{ prod_MWh_an }{  numeric }
  \item{ part_EnR }{  numeric }
  \item{ NOM_DEPCOM }{  character }
  \item{ EPCI }{  character }
  \item{ NOM_EPCI }{  character }
  \item{ DEP }{  character }
  \item{ NOM_DEP }{  character }
  \item{ REG }{  character }
  \item{ NOM_REG }{  character }
  \item{ DEPARTEMENTS_DE_L_EPCI }{  list }
  \item{ precision_geo }{  character }
  \item{ geometry }{  sfc_POINT,sfc }
}
}
\source{
DREAL PdL - TEO
}
\usage{
inst_reg
}
\description{
Un dataset rassemblant toutes les installations de production d'electricite renouvelable.
}
\keyword{datasets}
