% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_df.R
\docType{data}
\name{lib_filiere}
\alias{lib_filiere}
\title{lib_filiere}
\format{
A data frame with 12 rows and 7 variables:
\describe{
  \item{ Filiere_de_production }{ libelles sans accent des filieres visualisees, character }
  \item{ filiere }{  codes des filiere visualisee, character }
  \item{ fil_enedis_2017 }{ libelles des filieres Enedis, character }
  \item{ fil_registre }{ codes de filiere registre electricite, character }
  \item{ typo }{ libelles des filieres fines du registre electricite, factor }
  \item{ code_typo }{ codes des filieres fines du registre electricite, factor }
  \item{ filiere_registre }{ libelles des filiere du registre electricite, character }
}
}
\source{
DREAL PdL - TEO
}
\usage{
lib_filiere
}
\description{
Nomenclature des filieres electriques et leur correspondance, source par source.
}
\keyword{datasets}
