% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_values.R
\docType{data}
\name{liste_dep}
\alias{liste_dep}
\title{liste_dep}
\format{
Un vecteur nomme, au format texte
}
\source{
DREAL PdL - TEO
}
\usage{
liste_dep
}
\description{
Un vecteur nommé des départements de la région.
}
\keyword{datasets}
