% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_df.R
\docType{data}
\name{liste_zone_complete}
\alias{liste_zone_complete}
\title{liste_zone_complete}
\format{
A data frame with 78 rows and 3 variables:
\describe{
  \item{ TypeZone }{  Epci, Départements ou Régions, factor }
  \item{ Zone }{  le libelle du territoire, factor }
  \item{ CodeZone }{  le code du territoire, factor }
}
}
\source{
DREAL PdL - TEO
}
\usage{
liste_zone_complete
}
\description{
Un dataframe des territoires visualisables dans l'application : les 71 EPCI de la region + l'Ile d'Yeu, les 5 departements et la region.
}
\keyword{datasets}
