% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_df.R
\docType{data}
\name{liste_zones}
\alias{liste_zones}
\title{liste_zones}
\format{
Une liste nommee de 78 elements
}
\source{
DREAL PdL - TEO
}
\usage{
liste_zones
}
\description{
Une liste nommee des 78 territoires visualisables dans l'application. 
Chaque element est nomme par un libelle de territoire et contient le code geographique du territoire.
}
\keyword{datasets}
