% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_df.R
\docType{data}
\name{liste_zones_dep}
\alias{liste_zones_dep}
\title{liste_zones_dep}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{ dep }{  code departement, character }
  \item{ data }{  listes nommees des codes territoires relevant du departement, character }
}
}
\source{
DREAL PdL - TEO
}
\usage{
liste_zones_dep
}
\description{
Un dataframe rassemblant cinq listes nommees departementales. 
Chaque element d'une liste departementale est nomme par un libelle de territoire et contient le code geographique du territoire.
}
\keyword{datasets}
