% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cartes_inst_elec.R
\name{mes_instal}
\alias{mes_instal}
\title{Selection des installations d'une filiere electrique et d'un territoire avec jitter des coord geo et passage en CRS 4326}
\usage{
mes_instal(fil, code_ter, typ_exclus = "xxxxxxxxx")
}
\arguments{
\item{fil}{un motif de texte a retrouver dans le code typo des installations electrique qui permet d'identifier une unique filiere}

\item{code_ter}{le code geo du territoire a visualiser, au format texte}

\item{typ_exclus}{un motif de texte a retrouver dans le code typo des installations electrique pour exclure ce type d'installation de la liste}
}
\value{
un tibble sf
}
\description{
Selection des installations d'une filiere et d'un territoire avec jitter des coord geo et passage en CRS 4326
}
\examples{
mes_instal(fil = "eol", code_ter = "200060010")
}
