% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_df.R
\docType{data}
\name{metadatas}
\alias{metadatas}
\title{metadatas}
\format{
A data frame with 11 rows and 4 variables:
\describe{
  \item{ dataset }{  libelle du dataset, character }
  \item{ millesime }{ millesimes utilises, character }
  \item{ producteur }{ Organisme producteur du jeu de donnees, character }
  \item{ date_publication }{ Date de publication du datset, Date }
}
}
\source{
DREAL PdL - TEO
}
\usage{
metadatas
}
\description{
Metadonnees des datasets opendata collectes
}
\keyword{datasets}
