% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_values.R
\docType{data}
\name{mil_gaz}
\alias{mil_gaz}
\title{mil_gaz}
\format{
une valeur numerique
}
\source{
DREAL PdL - TEO
}
\usage{
mil_gaz
}
\description{
Le millesime des donnees gaz et biogaz.
}
\keyword{datasets}
