% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_df.R
\docType{data}
\name{obj_reg}
\alias{obj_reg}
\title{obj_reg}
\format{
A data frame with 9 rows and 8 variables:
\describe{
  \item{ TypeZone }{  Type de Zone, factor }
  \item{ CodeZone }{  Code de la Zone, factor }
  \item{ Filiere.de.production }{Filiere de production, factor }
  \item{ annee }{ les annees : 2014, annee d'adaption du SRCAE, 2020, 2050, character }
  \item{ indicateur }{ l'indicateur concerne : la puissance raccordee en kW, factor }
  \item{ valeur }{ valeur de l'indicateir, numeric }
  \item{ estimation }{ la valeur a t elle ete recalculee pour l'application, logical }
  \item{ type_estimation }{ origine de la valeur, character }
}
}
\source{
DREAL PdL - TEO
}
\usage{
obj_reg
}
\description{
Un dataset qui rassemble les objectifs regionaux de developpement des filieres par epoque.
}
\keyword{datasets}
