% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(onStart = NULL, options = list(), enableBookmarking = NULL, ...)
}
\arguments{
\item{onStart}{NULL, A function that will be called before the app is actually run.}

\item{options}{List, Named options that should be passed to the runApp call (these can be any of the following: "port", "launch.browser", "host", "quiet", "display.mode" and "test.mode"). You can also specify width and height parameters which provide a hint to the embedding environment about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of "url", "server", or "disable". The default value, NULL, will respect the setting from any previous calls to enableBookmarking(). See enableBookmarking() for more information on bookmarking your app.}

\item{...}{arguments to pass to golem_opts}
}
\description{
Run the Shiny Application
}
