% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tab_inst_biogaz.R
\name{tab_inst_biogaz}
\alias{tab_inst_biogaz}
\title{tab_inst_biogaz, filtre et mise en forme de la table des points d'injection biomethane}
\usage{
tab_inst_biogaz(ter = "49")
}
\arguments{
\item{ter}{le territoire a visualiser}
}
\value{
un dataframe
}
\description{
tab_inst_biogaz, filtre et mise en forme de la table des points d'injection biomethane
}
\examples{
tab_inst_biogaz("49")
}
