% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tab_inst_elec.R
\name{tab_inst_elec}
\alias{tab_inst_elec}
\title{tab_inst_elec, filtre et mise en forme de la table des installations de production electrique ENRR}
\usage{
tab_inst_elec(ter)
}
\arguments{
\item{ter}{le territoire a visualiser}
}
\value{
un dataframe
}
\description{
tab_inst_elec, filtre et mise en forme de la table des installations de production electrique ENRR
}
\examples{
tab_inst_elec("49")

}
