% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_themes_ggplot.R
\docType{data}
\name{theme_TEO}
\alias{theme_TEO}
\title{theme_TEO}
\format{
A data frame with  rows and  variables:
}
\source{
DREAL PdL
}
\usage{
theme_TEO
}
\description{
Le theme ggplot par defaut pour les graphiques de l'application.
}
\keyword{datasets}
