% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_themes_ggplot.R
\docType{data}
\name{theme_TEO_carto}
\alias{theme_TEO_carto}
\title{theme_TEO_carto}
\format{
theme ggplot
}
\source{
DREAL PdL
}
\usage{
theme_TEO_carto
}
\description{
Le theme ggplot par defaut pour les cartes de l'application.
}
\keyword{datasets}
