% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_df.R
\docType{data}
\name{typo_registre}
\alias{typo_registre}
\title{typo_registre}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{ code_typo }{ codes des filieres fines du registre electricite, character }
  \item{ typo }{  libelles des filieres fines du registre electricite, factor }
}
}
\source{
DREAL PdL - TEO
}
\usage{
typo_registre
}
\description{
Description.
}
\keyword{datasets}
